+++ src/rubric/criterion.rs
@@ -35,7 +35,7 @@ pub struct Criterion {
     ///
     /// Can be negative if you wish to subtract points. Be sure to get your logic right.
     /// This value is added to the submission grade *if the test returns true*.
-    pub worth: i16,
+    pub worth: isize,
     /// An index to sort by when running.
     ///
     /// Lowest first. Defaults to 100.
